Add assert.h
This implements assert(), a run-time assert function. This is similar to
the c standard library version of assert(), but is defined as weak. This
means that the assert implementation may be overruled to a
project-specific version.

For example: a unit test may overrule assert to fail the test instead
of hanging the program, while a firmware may choose to reboot if an
assert fails.