ringbuffer: refactor internal read/write pointers
The internal pointers (read and write) are now represented as offsets from
the start of the buffer. This makes the code a bit easier to understand.

The offset and wrap bit are packed together in a RingbufferIndex object.
It holds an offset that can be no larger than size_t - 1 bits.

Om a 32-bit mcu platform ,that is a 31-bits unsigned offset, plus a wrap
bit. By packing these together, they can be read and written as one 32-bit
word. This should guarantee that the offset + wrap bit are always in a
consistent state, even when viewed from a different thread / context /
cpu core.