Add Retry Ringbuffer
Very specific ringbuffer implementation that allows retries for reads. A
retry means that the ringbuffer is not advanced. The read element will
remain on the ringbuffer for a future read.

This ringbuffer automatically wraps, which means that the oldest element
will be overwritten when the buffer is full.

It uses the ringbuffer from this repo as underlying data structure.